requirements.txt
# Streamlit for frontend
streamlit
scikit-learn

# Firebase authentication and database
pyrebase4
firebase-admin
google-cloud-firestore

# Google Cloud dependencies (to replace deprecated gcloud)
google-cloud-storage

# Data handling and analysis
pandas

# Plotting
plotly

# HTTP requests
requests

# For working with time and threading
setuptools